adl/requirements.txt
wagtail==7.0.1
dj-database-url==2.2.0
django-timescaledb==0.2.13
django-environ==0.11.2
psycopg2==2.9.9
django-dbbackup==4.3.0
django-countries==7.6.1
django-countries-geoextent>=0.0.1
django-cleanup==8.1.0
pyoscar==0.8.0
django-timezone-field>=6.1.0
django-redis==5.4.0
celery-singleton==0.3.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
channels
daphne
minio==7.2.7
gunicorn==23.0.0
uvicorn[standard]==0.30.3
wagtail-font-awesome-svg==1.0.1
wagtail-icon-chooser>=0.3.0
pint==0.24.4
pandas==2.2.3
python-dateutil==2.9.0.post0
wagtailgeowidget==8.2.1
wagtail-modeladmin==2.0.0
django-polymorphic==3.1.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt>=5.5.0
django-vue-utils==0.1.7
django-deep-translator==1.4.0
drf-spectacular
more-itertools==10.7.0
